+++ components/content/SiteLogo.vue
@@ -1,5 +1,5 @@
 <template>
-  <nuxt-link to='/' class='text-2xl font-bold'><pre>&lt;/&gt;</pre></nuxt-link>
+  <nuxt-link to='/' class='text-2xl font-bold dark:text-gray-200 text-gray-700'><pre>&lt;/&gt;</pre></nuxt-link>
 </template>
 
 <script>