+++ tailwind.config.js
@@ -1,17 +1,15 @@
-const colors = require('tailwindcss/colors')
+import colors from 'tailwindcss/colors';
 
 module.exports = {
   darkMode: 'class',
-  plugins: [
-    require('@tailwindcss/typography')
-  ],
+  plugins: [require('@tailwindcss/typography')],
   content: ['content/**/*.{md,yml,json,json5,csv}'],
   theme: {
     extend: {
       colors: {
         // Customize the feeling of your site
-        gray: colors.stone
-      }
-    }
-  }
-}
+        gray: colors.stone,
+      },
+    },
+  },
+};